Only initialize RepeaterBook State ID if record has a RepeaterBook ID
The previous code was overzealous and initialized all missing
RepeaterBook State ID fields to Washington (53). This had the
unintended consequence of initializing the field for all repeaters
that are not in RepeaterBook (including those outside of Washington).

This change resets the state id for all repeaters w/o RepeterBook IDs
to null and updates the initialization code to:

- Prefer any existing RepeaterBook State ID

- Only initialize the state id to Washington (53) if the record has a
  RepeaterBook ID.